v45 полностью рабочий билд без интерфейса/requirements.txt
absl-py==2.3.0
aiodns==3.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiomysql==0.2.0
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.3.0
autopep8==2.3.2
bcrypt==3.2.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
ccxt==4.4.89
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.4
cssselect==1.3.0
cycler==0.12.1
dash==3.0.4
dateparser==1.2.1
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.104.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.25.2
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.3.2
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.4.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.4
namex==0.1.0
narwhals==1.42.1
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.16.0
optuna==4.4.0
packaging==25.0
pandas==2.3.0
pandas_ta==0.3.14b0
passlib==1.7.4
pillow==11.2.1
plotly==6.1.2
praw==7.8.1
prawcore==2.4.0
propcache==0.3.2
protobuf==5.29.5
psutil==5.9.6
pyasn1==0.6.1
pycares==4.9.0
pycodestyle==2.13.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.1
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.2.3
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.13.0
python-telegram-bot==20.7
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
retrying==1.3.4
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.15.3
sgmllib3k==1.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.23
starlette==0.27.0
structlog==23.2.0
sympy==1.14.0
ta==0.10.2
ta-lib==0.6.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
textblob==0.19.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
triton==3.3.1
tweepy==4.15.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.4.0
uvicorn==0.24.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.6
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.0.2
yarl==1.20.1
zipp==3.23.0